chat_bot_final/backend/requirements.txt
accelerate
aioboto3
aiohttp

# database packages
graspologic
hnswlib
nano-vectordb
neo4j
networkx
ollama
openai
oracledb
psycopg[binary,pool]
pymilvus
pymongo
pymysql
pyvis
# lmdeploy[all]
sqlalchemy
tenacity


# LLM packages
tiktoken
torch
transformers
xxhash
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langchain-tavily
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk

google-ai-generativelanguage
google-api-core
google-auth
google-genai
googleapis-common-protos

python-dotenv
pydantic
pydantic-settings
pydantic_core
fastapi
uvicorn
motor